refactor: Extract Navbar component and update Dashboard/AboutUs screens
