+++ Frontend/src/pages/AboutUs.jsx
@@ -13,8 +13,6 @@ import {
   Mail,
   Linkedin,
   Github,
-  Menu,
-  X,
   ArrowRight,
   MessageSquare,
   BarChart,
@@ -22,49 +20,47 @@ import {
   Lightbulb,
   Star,
 } from "lucide-react";
+import Navbar from "../components/shared/Navbar";
 import { useState } from "react";
 import Lenis from "@studio-freight/lenis";
 
 export default function AboutUs() {
   const navigate = useNavigate();
-  const { token } = useSelector((state) => state.auth);
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
+
   const lenisRef = useRef(null);
   const [contributors, setContributors] = React.useState([]);
   const [loading, setLoading] = React.useState(true);
   const [error, setError] = React.useState(null);
-  
-    // Initialize Lenis smooth scroll
-    useEffect(() => {
-      const lenis = new Lenis({
-        duration: 1.2,
-        easing: (t) => Math.min(1, 1.001 - Math.pow(2, -10 * t)),
-        orientation: "vertical",
-        gestureOrientation: "vertical",
-        smoothWheel: true,
-        wheelMultiplier: 1,
-        smoothTouch: false,
-        touchMultiplier: 2,
-        infinite: false,
-      });
-  
-      lenisRef.current = lenis;
-  
-      function raf(time) {
-        lenis.raf(time);
-        requestAnimationFrame(raf);
-      }
-  
-      requestAnimationFrame(raf);
-  
-      return () => {
-        lenis.destroy();
-      };
-    }, []);
 
-  const closeMobileMenu = () => {
-    setMobileMenuOpen(false);
-  };
+  // Initialize Lenis smooth scroll
+  useEffect(() => {
+    const lenis = new Lenis({
+      duration: 1.2,
+      easing: (t) => Math.min(1, 1.001 - Math.pow(2, -10 * t)),
+      orientation: "vertical",
+      gestureOrientation: "vertical",
+      smoothWheel: true,
+      wheelMultiplier: 1,
+      smoothTouch: false,
+      touchMultiplier: 2,
+      infinite: false,
+    });
+
+    lenisRef.current = lenis;
+
+    function raf(time) {
+      lenis.raf(time);
+      requestAnimationFrame(raf);
+    }
+
+    requestAnimationFrame(raf);
+
+    return () => {
+      lenis.destroy();
+    };
+  }, []);
+
+
   const itemVariants = {
     hidden: { y: 0, opacity: 1 },
     visible: {
@@ -77,7 +73,7 @@ export default function AboutUs() {
       },
     },
   };
-   const containerVariants = {
+  const containerVariants = {
     hidden: { opacity: 1 },
     visible: {
       opacity: 1,
@@ -181,7 +177,7 @@ export default function AboutUs() {
       },
       {
         name: "User-Centric",
-        role: "Design Philosophy", 
+        role: "Design Philosophy",
         description: "Every feature designed with user experience and accessibility in mind",
         icon: Heart,
       },
@@ -191,171 +187,7 @@ export default function AboutUs() {
   return (
     <div className="bg-white text-gray-900">
       {/* Navbar */}
-      <nav className="fixed top-6 left-1/2 transform -translate-x-1/2 w-[95%] max-w-7xl bg-white/95 backdrop-blur-md rounded-full shadow-lg z-50 border border-gray-200">
-        <div className="px-4 md:px-8 py-4 flex items-center justify-between">
-          <Link to="/" className="text-xl md:text-2xl font-bold">
-            <span className="text-gray-900">Interv</span>
-            <span className="text-emerald-500">yo</span>
-          </Link>
-
-          {/* Desktop Navigation */}
-          <div className="hidden lg:flex items-center gap-8">
-            <a
-              href="/#features"
-              className="text-gray-600 hover:text-gray-900 font-medium transition-colors cursor-pointer"
-            >
-              Features
-            </a>
-            <Link
-              to="/about"
-              className="text-emerald-500 font-medium transition-colors"
-            >
-              About
-            </Link>
-            <a
-              href="/#how-it-works"
-              className="text-gray-600 hover:text-gray-900 font-medium transition-colors cursor-pointer"
-            >
-              How it Works
-            </a>
-            <a
-              href="/#pricing"
-              className="text-gray-600 hover:text-gray-900 font-medium transition-colors cursor-pointer"
-            >
-              Pricing
-            </a>
-            <a
-              href="/#faq"
-              className="text-gray-600 hover:text-gray-900 font-medium transition-colors cursor-pointer"
-            >
-              FAQ
-            </a>
-            <Link
-              to="/contact"
-              className="text-gray-600 hover:text-gray-900 font-medium transition-colors"
-            >
-              Contact
-            </Link>
-          </div>
-
-          {/* Desktop Buttons */}
-          <div className="hidden lg:flex items-center gap-4">
-            {token ? (
-              <button
-                onClick={() => navigate("/dashboard")}
-                className="px-6 py-2.5 bg-black text-white rounded-lg hover:bg-gray-800 font-semibold shadow-lg transition-all"
-              >
-                Dashboard
-              </button>
-            ) : (
-              <>
-                <Link
-                  to="/login"
-                  className="px-4 py-2 bg-black text-white rounded-lg hover:bg-gray-800 font-semibold shadow-lg transition-all text-sm"
-                >
-                  Sign In
-                </Link>
-                <Link
-                  to="/register"
-                  className="px-4 py-2 bg-emerald-500 text-white rounded-lg hover:bg-emerald-600 font-semibold shadow-lg transition-all text-sm"
-                >
-                  Get Started
-                </Link>
-              </>
-            )}
-          </div>
-
-          {/* Mobile Hamburger Button */}
-          <button
-            onClick={() => setMobileMenuOpen(!mobileMenuOpen)}
-            className="lg:hidden p-2 text-gray-900 hover:bg-gray-100 rounded-lg transition-colors"
-            aria-label="Toggle menu"
-          >
-            {mobileMenuOpen ? <X size={24} /> : <Menu size={24} />}
-          </button>
-        </div>
-
-        {/* Mobile Menu */}
-        {mobileMenuOpen && (
-          <div className="lg:hidden absolute top-full left-0 right-0 mt-2 bg-white backdrop-blur-md rounded-2xl shadow-xl border border-gray-200 mx-2 overflow-hidden">
-            <div className="p-6 space-y-4">
-              <a
-                href="/#features"
-                className="block text-gray-600 hover:text-gray-900 font-medium py-3 px-4 rounded-lg hover:bg-gray-50 transition-colors"
-                onClick={closeMobileMenu}
-              >
-                Features
-              </a>
-              <a
-                href="/#how-it-works"
-                className="block text-gray-600 hover:text-gray-900 font-medium py-3 px-4 rounded-lg hover:bg-gray-50 transition-colors"
-                onClick={closeMobileMenu}
-              >
-                How it Works
-              </a>
-              <a
-                href="/#pricing"
-                className="block text-gray-600 hover:text-gray-900 font-medium py-3 px-4 rounded-lg hover:bg-gray-50 transition-colors"
-                onClick={closeMobileMenu}
-              >
-                Pricing
-              </a>
-              <a
-                href="/#faq"
-                className="block text-gray-600 hover:text-gray-900 font-medium py-3 px-4 rounded-lg hover:bg-gray-50 transition-colors"
-                onClick={closeMobileMenu}
-              >
-                FAQ
-              </a>
-              <Link
-                to="/contact"
-                onClick={closeMobileMenu}
-                className="block text-gray-600 hover:text-gray-900 font-medium py-3 px-4 rounded-lg hover:bg-gray-50 transition-colors"
-              >
-                Contact
-              </Link>
-              <Link
-                to="/about"
-                onClick={closeMobileMenu}
-                className="block text-emerald-500 font-medium py-3 px-4 rounded-lg bg-emerald-50"
-              >
-                About
-              </Link>
-
-              <div className="pt-4 border-t border-gray-200 space-y-3">
-                {token ? (
-                  <button
-                    onClick={() => {
-                      navigate("/dashboard");
-                      closeMobileMenu();
-                    }}
-                    className="w-full px-6 py-3 bg-black text-white rounded-lg hover:bg-gray-800 font-semibold shadow-lg transition-all"
-                  >
-                    Dashboard
-                  </button>
-                ) : (
-                  <>
-                    <Link
-                      to="/login"
-                      onClick={closeMobileMenu}
-                      className="block w-full px-6 py-3 bg-black text-white rounded-lg hover:bg-gray-800 font-semibold shadow-lg transition-all text-center"
-                    >
-                      Sign In
-                    </Link>
-                    <Link
-                      to="/register"
-                      onClick={closeMobileMenu}
-                      className="block w-full px-6 py-3 bg-emerald-500 text-white rounded-lg hover:bg-emerald-600 font-semibold shadow-lg transition-all text-center"
-                    >
-                      Get Started
-                    </Link>
-                  </>
-                )}
-              </div>
-            </div>
-          </div>
-        )}
-      </nav>
+      <Navbar variant="public" />
 
       {/* Hero Section */}
       <section className="pt-32 pb-20 px-6 relative overflow-hidden bg-gray-950">
@@ -665,181 +497,181 @@ export default function AboutUs() {
             ))}
           </div> */}
           {/* Creator Section */}
-        <AnimatedSection
-          className={`rounded-3xl shadow-2xl p-8 md:p-12 mb-12 border backdrop-blur-sm parallax-element bg-gray-900 border-gray-700/50
+          <AnimatedSection
+            className={`rounded-3xl shadow-2xl p-8 md:p-12 mb-12 border backdrop-blur-sm parallax-element bg-gray-900 border-gray-700/50
               `}
-        >
-          <div className="text-center mb-8">
-            <motion.div
-              className="flex justify-center mb-6"
-              whileHover={{ scale: 1.1 }}
-              transition={{ type: "spring", stiffness: 300 }}
-            >
+          >
+            <div className="text-center mb-8">
               <motion.div
-                className={`relative p-2 rounded-2xl bg-blue-500/20`}
-                animate={{
-                  boxShadow: [
-                    "0 0 20px rgba(59, 130, 246, 0.3)",
-                    "0 0 40px rgba(59, 130, 246, 0.6)",
-                    "0 0 20px rgba(59, 130, 246, 0.3)",
-                  ],
-                }}
-                transition={{ duration: 2, repeat: Infinity }}
+                className="flex justify-center mb-6"
+                whileHover={{ scale: 1.1 }}
+                transition={{ type: "spring", stiffness: 300 }}
               >
-                <img
-                  src={team.creator.avatar}
-                  alt={team.creator.name}
-                  className="w-32 h-32 rounded-xl object-cover"
-                  onError={(e) => {
-                    e.target.src = `https://ui-avatars.com/api/?name=${team.creator.name}&background=3b82f6&color=ffffff&size=128`;
-                  }}
-                />
                 <motion.div
-                  className="absolute -top-2 -right-2"
-                  animate={{ rotate: 360 }}
-                  transition={{ duration: 4, repeat: Infinity, ease: "linear" }}
+                  className={`relative p-2 rounded-2xl bg-blue-500/20`}
+                  animate={{
+                    boxShadow: [
+                      "0 0 20px rgba(59, 130, 246, 0.3)",
+                      "0 0 40px rgba(59, 130, 246, 0.6)",
+                      "0 0 20px rgba(59, 130, 246, 0.3)",
+                    ],
+                  }}
+                  transition={{ duration: 2, repeat: Infinity }}
                 >
-                  <Star className="w-8 h-8 text-yellow-500 fill-yellow-500" />
+                  <img
+                    src={team.creator.avatar}
+                    alt={team.creator.name}
+                    className="w-32 h-32 rounded-xl object-cover"
+                    onError={(e) => {
+                      e.target.src = `https://ui-avatars.com/api/?name=${team.creator.name}&background=3b82f6&color=ffffff&size=128`;
+                    }}
+                  />
+                  <motion.div
+                    className="absolute -top-2 -right-2"
+                    animate={{ rotate: 360 }}
+                    transition={{ duration: 4, repeat: Infinity, ease: "linear" }}
+                  >
+                    <Star className="w-8 h-8 text-yellow-500 fill-yellow-500" />
+                  </motion.div>
                 </motion.div>
               </motion.div>
-            </motion.div>
 
-            <motion.h2
-              className={`text-4xl font-bold mb-4 text-white`}
-              variants={itemVariants}
-            >
-              Meet the <span className="text-emerald-500">Creator</span>
-            </motion.h2>
-
-            <motion.div
-              className="bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent"
-              variants={itemVariants}
-            >
-              <h3 className="text-2xl font-bold mb-2 text-emerald-500">{team.creator.name}</h3>
-              <p className="text-lg font-semibold text-emerald-500">{team.creator.role}</p>
-            </motion.div>
-
-            <motion.p
-              className={`text-lg leading-relaxed mt-4 max-w-2xl mx-auto text-gray-300`}
-              variants={itemVariants}
-            >
-              {team.creator.description}
-            </motion.p>
-
-            <motion.div
-              className="flex justify-center gap-4 mt-6"
-              variants={itemVariants}
-            >
-              <motion.a
-                href={team.creator.github}
-                target="_blank"
-                rel="noopener noreferrer"
-                className={`p-3 rounded-xl transition-all duration-300 bg-gray-700 hover:bg-gray-600 text-white"
-                    `}
-                whileHover={{ scale: 1.1, y: -2 }}
-                whileTap={{ scale: 0.95 }}
+              <motion.h2
+                className={`text-4xl font-bold mb-4 text-white`}
+                variants={itemVariants}
               >
-                <Github className="w-6 h-6" />
-              </motion.a>
-              <motion.a
-                href={`mailto:${team.creator.email}`}
-                className={`p-3 rounded-xl transition-all duration-300 bg-blue-600 hover:bg-blue-500 text-white"
-                    `}
-                whileHover={{ scale: 1.1, y: -2 }}
-                whileTap={{ scale: 0.95 }}
-              >
-                <Mail className="w-6 h-6" />
-              </motion.a>
-            </motion.div>
+                Meet the <span className="text-emerald-500">Creator</span>
+              </motion.h2>
 
-            <motion.div
-              className="mt-6"
-              variants={itemVariants}
-            >
-              <p className={`text-sm text-gray-400 mb-3`}>
-                Tech Stack Expertise:
-              </p>
-              <div className="flex flex-wrap justify-center gap-2">
-                {team.creator.skills.map((skill, idx) => (
-                  <motion.span
-                    key={idx}
-                    className={`px-3 py-1 rounded-full text-sm font-medium bg-blue-500/20 text-blue-400 border border-blue-500/30"
-                        `}
-                    whileHover={{ scale: 1.05 }}
-                    initial={{ opacity: 0, y: 10 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    transition={{ delay: idx * 0.1 }}
-                  >
-                    {skill}
-                  </motion.span>
-                ))}
-              </div>
-            </motion.div>
-          </div>
-        </AnimatedSection>
-          {/* Contributors Section */}
-        <div
-          className={`rounded-3xl shadow-2xl p-8 md:p-12 mb-12 border backdrop-blur-sm bg-gray-950 border-gray-700/50`}
-        >
-          <div className="flex items-center gap-4 mb-6">
-            <div
-              className={`p-3 rounded-2xl bg-purple-500/20`}
-            >
-              <Github
-                className={`w-8 h-8 text-purple-400`}
-              />
-            </div>
-            <h2
-              className={`text-3xl font-bold text-white`}
-            >
-              Project Contributors
-            </h2>
-          </div>
-          
-          {loading ? (
-            <div className="flex justify-center items-center py-8">
-              <div className={`text-lg text-gray-400`}>
-                Loading contributors...
-              </div>
-            </div>
-          ) : error ? (
-            <div className={`text-center p-6 rounded-xl bg-red-900/30`}>
-              <p className={`ext-red-400`}>
-                Error loading contributors: {error}
-              </p>
-              <p className={`text-gray-400`}>
-                Visit our <a href="https://github.com/santanu-atta03/Intervyo" target="_blank" rel="noopener noreferrer" className="underline">GitHub repository</a> to see contributors.
-              </p>
-            </div>
-          ) : (
-            <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-6 gap-4">
-              {contributors.map((contributor) => (
-                <a
-                  key={contributor.id}
-                  href={contributor.html_url}
+              <motion.div
+                className="bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent"
+                variants={itemVariants}
+              >
+                <h3 className="text-2xl font-bold mb-2 text-emerald-500">{team.creator.name}</h3>
+                <p className="text-lg font-semibold text-emerald-500">{team.creator.role}</p>
+              </motion.div>
+
+              <motion.p
+                className={`text-lg leading-relaxed mt-4 max-w-2xl mx-auto text-gray-300`}
+                variants={itemVariants}
+              >
+                {team.creator.description}
+              </motion.p>
+
+              <motion.div
+                className="flex justify-center gap-4 mt-6"
+                variants={itemVariants}
+              >
+                <motion.a
+                  href={team.creator.github}
                   target="_blank"
                   rel="noopener noreferrer"
-                  className={`block group p-4 rounded-xl transition-all duration-300 hover:scale-105 hover:shadow-lg hover:bg-gray-700/50 border border-gray-700/50`}
-                  title={`${contributor.login} (${contributor.contributions} contributions)`}
+                  className={`p-3 rounded-xl transition-all duration-300 bg-gray-700 hover:bg-gray-600 text-white"
+                    `}
+                  whileHover={{ scale: 1.1, y: -2 }}
+                  whileTap={{ scale: 0.95 }}
                 >
-                  <div className="flex flex-col items-center">
-                    <img
-                      src={contributor.avatar_url}
-                      alt={contributor.login}
-                      className="w-16 h-16 rounded-full mb-3 object-cover border-2 border-transparent group-hover:border-purple-500 transition-colors duration-300"
-                    />
-                    <h3 className={`font-semibold text-center truncate w-full text-white`}>
-                      {contributor.login}
-                    </h3>
-                    <p className={`text-sm mt-1 text-gray-400`}>
-                      {contributor.contributions} contributions
-                    </p>
-                  </div>
-                </a>
-              ))}
+                  <Github className="w-6 h-6" />
+                </motion.a>
+                <motion.a
+                  href={`mailto:${team.creator.email}`}
+                  className={`p-3 rounded-xl transition-all duration-300 bg-blue-600 hover:bg-blue-500 text-white"
+                    `}
+                  whileHover={{ scale: 1.1, y: -2 }}
+                  whileTap={{ scale: 0.95 }}
+                >
+                  <Mail className="w-6 h-6" />
+                </motion.a>
+              </motion.div>
+
+              <motion.div
+                className="mt-6"
+                variants={itemVariants}
+              >
+                <p className={`text-sm text-gray-400 mb-3`}>
+                  Tech Stack Expertise:
+                </p>
+                <div className="flex flex-wrap justify-center gap-2">
+                  {team.creator.skills.map((skill, idx) => (
+                    <motion.span
+                      key={idx}
+                      className={`px-3 py-1 rounded-full text-sm font-medium bg-blue-500/20 text-blue-400 border border-blue-500/30"
+                        `}
+                      whileHover={{ scale: 1.05 }}
+                      initial={{ opacity: 0, y: 10 }}
+                      animate={{ opacity: 1, y: 0 }}
+                      transition={{ delay: idx * 0.1 }}
+                    >
+                      {skill}
+                    </motion.span>
+                  ))}
+                </div>
+              </motion.div>
             </div>
-          )}
-        </div>
+          </AnimatedSection>
+          {/* Contributors Section */}
+          <div
+            className={`rounded-3xl shadow-2xl p-8 md:p-12 mb-12 border backdrop-blur-sm bg-gray-950 border-gray-700/50`}
+          >
+            <div className="flex items-center gap-4 mb-6">
+              <div
+                className={`p-3 rounded-2xl bg-purple-500/20`}
+              >
+                <Github
+                  className={`w-8 h-8 text-purple-400`}
+                />
+              </div>
+              <h2
+                className={`text-3xl font-bold text-white`}
+              >
+                Project Contributors
+              </h2>
+            </div>
+
+            {loading ? (
+              <div className="flex justify-center items-center py-8">
+                <div className={`text-lg text-gray-400`}>
+                  Loading contributors...
+                </div>
+              </div>
+            ) : error ? (
+              <div className={`text-center p-6 rounded-xl bg-red-900/30`}>
+                <p className={`ext-red-400`}>
+                  Error loading contributors: {error}
+                </p>
+                <p className={`text-gray-400`}>
+                  Visit our <a href="https://github.com/santanu-atta03/Intervyo" target="_blank" rel="noopener noreferrer" className="underline">GitHub repository</a> to see contributors.
+                </p>
+              </div>
+            ) : (
+              <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-6 gap-4">
+                {contributors.map((contributor) => (
+                  <a
+                    key={contributor.id}
+                    href={contributor.html_url}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className={`block group p-4 rounded-xl transition-all duration-300 hover:scale-105 hover:shadow-lg hover:bg-gray-700/50 border border-gray-700/50`}
+                    title={`${contributor.login} (${contributor.contributions} contributions)`}
+                  >
+                    <div className="flex flex-col items-center">
+                      <img
+                        src={contributor.avatar_url}
+                        alt={contributor.login}
+                        className="w-16 h-16 rounded-full mb-3 object-cover border-2 border-transparent group-hover:border-purple-500 transition-colors duration-300"
+                      />
+                      <h3 className={`font-semibold text-center truncate w-full text-white`}>
+                        {contributor.login}
+                      </h3>
+                      <p className={`text-sm mt-1 text-gray-400`}>
+                        {contributor.contributions} contributions
+                      </p>
+                    </div>
+                  </a>
+                ))}
+              </div>
+            )}
+          </div>
 
           <p className="text-gray-500 text-center mt-12 max-w-2xl mx-auto">
             We're a growing team of passionate developers, designers, and AI
